feat(sitevision-scripts)!: don't add main.js in build step
Previously a main.js has been created during build time if it was missing in source.
Since Sitevision 2025.07.1 the main.js is no longer required for WebApps.

BREAKING CHANGE: If you're building for < Sitevision 2025.07.1.
You need to add an main.js to be able to deploy your app to Sitevision.